refactor: extract silenceUseClientWarning to common package
