+++ packages/common/package.json
@@ -6,5 +6,11 @@
   "exports": {
     ".": "./index.ts",
     "./refresh-runtime": "./refresh-runtime.js"
+  },
+  "peerDependencies": {
+    "vite": "^4.2.0 || ^5.0.0 || ^6.0.0"
+  },
+  "devDependencies": {
+    "vite": "^6.2.2"
   }
 }
