+++ packages/plugin-react-swc/package.json
@@ -38,7 +38,7 @@
   "devDependencies": {
     "@playwright/test": "^1.51.1",
     "@types/fs-extra": "^11.0.4",
-    "@types/node": "22.13.10",
+    "@types/node": "^22.13.15",
     "@vitejs/react-common": "workspace:*",
     "@vitejs/release-scripts": "^1.3.3",
     "esbuild": "^0.25.1",
